Move this typescript test in with our other tests
Remove an npm script & generated dir while we're at it
