+++ package.json
@@ -19,8 +19,8 @@
     "build_file": "dist/testdouble.js"
   },
   "scripts": {
-    "clean": "rm -rf generated dist lib coverage",
-    "postclean": "mkdirp generated dist",
+    "clean": "rm -rf dist lib coverage",
+    "postclean": "mkdirp dist",
     "compile:browser": "browserify src/index.js --standalone td --outfile $npm_package_config_build_file -p tsify -p headerify",
     "compile:node": "tsc --outDir lib",
     "precompile": "npm run clean",
@@ -28,12 +28,11 @@
     "cover": "nyc --reporter=lcov --reporter=text-summary --reporter=html npm test",
     "cover:report": "codeclimate-test-reporter < coverage/lcov.info",
     "style": "standard --fix",
-    "test": "teenytest --helper test/helper.js \"test/**/*.test.js\"",
-    "test:all": "npm test && npm run test:example && npm run test:typescript",
-    "test:unit": "teenytest --helper test/helper.js \"test/unit/**/*.test.js\"",
-    "test:safe": "teenytest --helper test/helper.js \"test/safe/**/*.test.js\"",
+    "test": "teenytest --helper test/helper.js \"test/**/*.test.{js,ts}\"",
+    "test:all": "npm test && npm run test:example",
+    "test:unit": "teenytest --helper test/helper.js \"test/unit/**/*.test.{js,ts}\"",
+    "test:safe": "teenytest --helper test/helper.js \"test/safe/**/*.test.{js,ts}\"",
     "test:ci": "npm run compile && npm run style && npm run test:all && echo \"All done!\"",
-    "test:typescript": "tsc --outDir generated/typescript -p test/typescript && node generated/typescript/test.js",
     "test:example": "./script/run-examples",
     "version:write": "echo \"export default '$npm_package_version'\" > src/version.js",
     "version:changelog": "if command -v github_changelog_generator &>/dev/null; then github_changelog_generator; git commit -m \"Changelog for $npm_package_version\" CHANGELOG.md; else echo Versioning requires you first run 'gem install github_changelog_generator' >&2; fi",
